+++ Calculate/Views/Operation/Index.cshtml
@@ -1,0 +1,99 @@
+﻿@{
+    ViewData["Title"] = "Operation Page";
+}
+@model IEnumerable<Calculate.Data.Models.Operation>
+
+<div class="app-content content">
+    <div class="content-overlay"></div>
+    <div class="content-wrapper">
+        <div class="content-header row">
+        </div>
+        <div class="content-body">
+            <div class="row">
+                <section id="file-export">
+                    <div class="row">
+                        <div class="col-12">
+                            <div class="card">
+                                <div class="card-header">
+                                    <h4 class="card-title">File export</h4>
+                                    <a class="heading-elements-toggle"><i class="la la-ellipsis-v font-medium-3"></i></a>
+                                    <div class="heading-elements">
+                                        <ul class="list-inline mb-0">
+                                            <li> <a class="btn btn-success btn-sm" href="/Operation/OperationCreate/"><i class="ft-plus white"></i> Ekle</a></li>
+                                            <li><a data-action="collapse"><i class="ft-minus"></i></a></li>
+                                            <li><a data-action="reload"><i class="ft-rotate-cw"></i></a></li>
+                                            <li><a data-action="expand"><i class="ft-maximize"></i></a></li>
+                                            <li><a data-action="close"><i class="ft-x"></i></a></li>
+                                        </ul>
+                                    </div>
+                                </div>
+                                <div class="card-content collapse show">
+                                    <div class="card-body card-dashboard dataTables_wrapper dt-bootstrap">
+                                        <p class="card-text">
+                                            Exporting data from a table can often be a key part of a complex
+                                            application. The Buttons extension for DataTables provides three plug-ins that provide
+                                            overlapping functionality for data export.
+                                        </p>
+                                        <table id="example" class="table table-striped table-bordered file-export">
+                                            <thead>
+                                                <tr>
+                                                    <th>İşlem No</th>
+                                                    <th>Hesap Adı</th>
+                                                    <th>Banka Adı</th>
+                                                    <th>İşlem Tipi</th>
+                                                    <th>Miktar</th>
+                                                    <th>Komisyon</th>
+                                                    <th>Güncelle</th>
+                                                    <th>Sil</th>
+                                                </tr>
+                                            </thead>
+                                            <tbody>
+                                                @foreach (var item in Model)
+                                                {
+                                                    <tr>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.ProcessNumber)
+                                                        </td>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.AccountId)
+                                                        </td>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.AccountDetailId)
+                                                        </td>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.ProcessTypeId)
+                                                        </td>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.Price)
+                                                        </td>
+                                                        <td>
+                                                            @Html.DisplayFor(modelItem => item.ProcessPrice)
+                                                        </td>
+                                                        <td><a class="btn btn-warning btn-sm" href="/Operation/OperationEdit/@item.Id"><i class="la la-pencil"></i> Güncelle</a></td>
+                                                        <td><a class="btn btn-danger btn-sm" href="/Operation/OperationDelete/@item.Id"><i class="la la-trash"></i> Sil</a></td>
+                                                    </tr>
+                                                }
+                                            </tbody>
+                                            <tfoot>
+                                                <tr>
+                                                    <th>İşlem No</th>
+                                                    <th>Hesap Adı</th>
+                                                    <th>Banka Adı</th>
+                                                    <th>İşlem Tipi</th>
+                                                    <th>Miktar</th>
+                                                    <th>Komisyon</th>
+                                                    <th>Güncelle</th>
+                                                    <th>Sil</th>
+                                                </tr>
+                                            </tfoot>
+                                        </table>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </section>
+            </div>
+        </div>
+    </div>
+</div>
