+++ Calculate/Views/Operation/OperationCreate.cshtml
@@ -64,7 +64,7 @@
             </div>
 
             <div class="form-actions text-center">
-                <a type="button" class="btn btn-warning mr-1" href="/Home/Index/">
+                <a type="button" class="btn btn-warning mr-1" href="/Operation/Index/">
                     <i class="ft-x"></i> Cancel
                 </a>
                 <button type="submit" class="btn btn-primary">
