+++ Calculate/Views/Shared/_Layout.cshtml
@@ -64,9 +64,9 @@
                                         <h6 class="font-weight-bold font-medium-2 ml-1">Admin Panel</h6>
                                         <ul class="row mt-2">
                                             <li class="col-6 col-xl-4">
-                                                <a class="text-center mb-2 mb-xl-3" href="../ecommerce-menu-template" target="_blank">
+                                                <a class="text-center mb-2 mb-xl-3" href="~/Operation/Index" target="_blank">
                                                     <i class="la la-shopping-cart font-large-1 mr-0"></i>
-                                                    <p class="font-medium-2 mt-25 mb-0">eCommerce</p>
+                                                    <p class="font-medium-2 mt-25 mb-0">İşlemler</p>
                                                 </a>
                                             </li>
                                             <li class="col-6 col-xl-4">
@@ -279,7 +279,7 @@
                     <span data-i18n="Admin Panels">Admin Panels</span><i class="material-icons nav-menu-icon" data-toggle="tooltip" data-placement="right" data-original-title="Admin Panels">more_horiz</i>
                 </li>
                 <li class=" nav-item">
-                    <a href="../material-ecommerce-menu-template" target="_blank"><i class="material-icons">add_shopping_cart</i><span class="menu-title" data-i18n="eCommerce">eCommerce</span></a>
+                    <a href="~/Operation/Index" target="_blank"><i class="material-icons">add_shopping_cart</i><span class="menu-title" data-i18n="eCommerce">İşlemler</span></a>
                 </li>
                 <li class=" nav-item">
                     <a href="../material-travel-menu-template" target="_blank"><i class="material-icons">call_merge</i><span class="menu-title" data-i18n="Travel &amp; Booking">Travel &amp; Booking</span></a>
